feat(messaging): move topic pages
Show subscribers first as that's probably the most useful/important thing for developers.

Move/rename details to settings to match other pages.
