+++ src/routes/console/project-[project]/messaging/topics/topic-[topic]/header.svelte
@@ -11,11 +11,6 @@
     const tabs = [
         {
             href: path,
-            title: 'Overview',
-            event: 'overview'
-        },
-        {
-            href: `${path}/subscribers`,
             title: 'Subscribers',
             event: 'subscribers'
         },
@@ -23,6 +18,11 @@
             href: `${path}/activity`,
             title: 'Activity',
             event: 'activity'
+        },
+        {
+            href: `${path}/settings`,
+            title: 'Settings',
+            event: 'settings'
         }
     ];
 </script>
